feat: Destroy localstorage
We introduce a `destroy` method in the storage interface, that will
be called after a `client.reset()`.
We implement this `destroy` in the localStorage platform, but it should
be implemented accordingly to deal with other storage context, such as
with indexedDB for workers or async-storage for react-native
