+++ packages/cozy-pouch-link/src/localStorage.js
@@ -14,6 +14,15 @@ export class PouchLocalStorage {
   }
 
   /**
+   * Destroy the storage data
+   *
+   * @returns {Promise<void>}
+   */
+  async destroy() {
+    this.storageEngine.destroy()
+  }
+
+  /**
    * Persist the last replicated doc id for a doctype
    *
    * @param {string} doctype - The replicated doctype
