+++ packages/cozy-pouch-link/src/replicateOnce.js
@@ -11,6 +11,7 @@ import {
   isDatabaseUnradableError
 } from './remote'
 import { startReplication } from './startReplication'
+import { allSettled } from './utils'
 
 /**
  * Process replication once for given PouchManager
@@ -159,45 +160,4 @@ export const replicateOnce = async pouchManager => {
   } catch (err) {
     pouchManager.handleReplicationError(err)
   }
-}
-
-/**
- * @typedef {object} FulfilledPromise
- * @property {'fulfilled'} status - The status of the promise
- * @property {undefined} reason - The Error rejected by the promise (undefined when fulfilled)
- * @property {any} value - The resolved value of the promise
- */
-
-/**
- * @typedef {object} RejectedPromise
- * @property {'rejected'} status - The status of the promise
- * @property {Error} reason - The Error rejected by the promise
- * @property {undefined} value - The resolved value of the promise (undefined when rejected)
- */
-
-/**
- * Takes an iterable of promises as input and returns a single Promise.
- * This returned promise fulfills when all of the input's promises settle (including
- * when an empty iterable is passed), with an array of objects that describe the
- * outcome of each promise.
- *
- * @param {Promise[]} promises - Promise to be awaited
- * @returns {Promise<(FulfilledPromise|RejectedPromise)[]>}
- */
-const allSettled = promises => {
-  return Promise.all(
-    promises.map(promise =>
-      promise
-        .then(value => /** @type {FulfilledPromise} */ ({
-          status: 'fulfilled',
-          value
-        }))
-        .catch((
-          /** @type {Error} */ reason
-        ) => /** @type {RejectedPromise} */ ({
-          status: 'rejected',
-          reason
-        }))
-    )
-  )
 }
